requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
python-dotenv>=1.0,<2.0
aiosqlite>=0.20.0,<0.21.0
psutil
matplotlib
